Fixup labels on Deployment and HPA
The `app.kubernetes.io/managed-by` label value referred to
`.Values.managed_by` variable which is undefined (possibly a typo for
`managed-by`) which caused an error when templating the chart.

Instead make the label refer to `.Release.Service` which is consistent
with the `Ingress` and `Service` resources and compliant with Helm
recommendations.